LVPN-8733: set MTU based on the route's MTU and header size for quench
Previously MTU was hardcoded to be 1500, but this would cause issues for
singular packets whose size would exceed the actual MTU.
